Add a generic loader for maven projects on jenkins.
This loader should be able to load any maven project on jenkins. It will read any  files from  for info on how to load the type defined in that file. To get the version number, it fetches the pom.xml file from a predefined path, given by the commit that started the build.